soc: qcom: smd: Add memcpy functions to handle 8-byte alignment
The standard memcpy implementation is not intended to be used for device
memory and hence does not handle the more restrictive device-memory
alignment constraints.  This may lead to a data abort when transferring
unaligned data to or from device memory when the transfer size is
greater than 1 byte.

Implement memcpy that will do byte-aligned copies until the address is
8-byte aligned, do multi-byte aligned copies, and then finish with
byte-aligned copies again as necessary.

Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>